\name{phylomap}
\alias{phylomap}
\docType{package}
\title{
Phylogenetic Stochastic Mapping
}
\description{
phylomap performs phylogenetic stochastic mapping, with matrix exponentiation, without matrix exponentiation, and without matrix exponentiation with sparsity.
}
\details{
\tabular{ll}{
Package: \tab phylomap\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-01-05\cr
License: GPL-3
}
phylomap has seven functions that return summary statisitics of each of N sampled substitution histories. The functions are sumstatEXP, sumstatMCMC, SPARSEsumstatMCMC, sumstatMCMCbf, sumstatMCMCks, sumstatMCMCmt, and sumstatMCMCksmt.
}
\author{
Jan Irvahn, Vladimir Minin

Maintainer: <vminin@uw.edu>

}
\references{

Irvahn, Jan, and Vladimir N. Minin. "Phylogenetic Stochastic Mapping Without Matrix Exponentiation." Journal of Computational Biology 21.9 (2014): 676-690.

}
