\name{phylomap}
\alias{phylomap}
\docType{package}
\title{
Phylogenetic Stochastic Mapping
}
\description{
phylomap performs phylogenetic stochastic mapping, with matrix exponentiation, without matrix exponentiation, and without matrix exponentiation with sparsity.
}
\details{
\tabular{ll}{
Package: \tab phylomap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-06\cr
License: GPL-3
}
phylomap has three functions that return the sufficient statisitics of each of N sampled substitution histories. The functions are sumstatEXP, sumstatMCMC, and SPARSEsumstatMCMC.
}
\author{
Jan Irvahn, Vladimir Minin

Maintainer: <vminin@uw.edu>

}
\references{

}
